Docker Containerization Cook Book
 ** This book is brief tutorial for beginners **

Docker is a container manager, which means that is able to create and execute containers that represent specific runtime environ-
ments for your software. In contrast with virtual machines like VirtualBox, Docker uses resource isolation features of the Linux
kernel to allow independent “containers” to run within a single Linux instance, avoiding the overhead of starting and maintaining
virtual machines. A computer with docker can run multiple containers at the same time.