The Linux Kernel Module Programming Guide
This guide is for anyone wishing to get started on kernel module programming.
The author takes a hands on approach starting  with writing the hello world program, and
quickly moves from there. Far from a boring text on programming.
This guide has a lively style that entertains while it educates.