The Linux Programming Interface A Linux and UNIX System Programming Handbook
This book is aimed primarily at the following audience:

 -Programmers and software designers building applications for Linux, other
UNIX systems, or other POSIX-conformant systems;

 -Programmers porting applications between Linux and other UNIX implemen-
tations or between Linux and other operating systems;

 -Instructors and advanced students teaching or learning Linux or UNIX system
programming; and

 -System managers and “power users” wishing to gain a greater understanding of
the Linux/UNIX programming interface and of how various pieces of system
software are implemented

This book assume you have some prior programming experience, but no previous system
programming experience is required.It also assume you have a reading knowledge
of the C programming language, and know how to use the shell and common Linux
or Unix commands.

If you are new to Linux or Unix, you will find it helpful to read the programmer-oriented
reqview of fundamental concepts of Linux and Unix Systen in the Chapter 2.