Invent Your Own Computer Games with Python 2nd Edition By Albert Sweigart
This is a free eBook that teaches you how to program in the Python Programming language.
Each chapter gives you the comple source code for a new game, and then teach the programming concepts
from the example, It was written to be understandable by kids as young as 10  to 12 years old,it is great for
anyone of any age who has never programmed before.

This book is for:
 - Complete beginners who wants to teach themselves computer programming, even if they have no prev. experience.
 - Kids and teenagers who want to learn computer programming by creating games.
 - Adults and teachers who wish to teach others programming
 - Aynone, young or old, who wants to learn to program by learning a professional programming language.