Learning Java by Building Android Games
If you are completely new to any one of Java, Android, or game programming, and
are aiming at publishing Android games for fun or for a business purpose, but you
have no clue as to where to start, then this book is for you.
This book also acts as a refresher for those who already have experience in Java on other platforms.
Android is the fastest growing operating system, and Android devices can empower,
entertain, and educate the planet.
Android uses one of the most popular programming languages—Java, which is a high-performance, secure, and object-oriented language.
In each chapter, we will build on what you learned from the previous chapter,
steadily gaining an understanding of more advanced Java concepts and putting
them to work by building Android games.

By the end of this book, we will have made four complete games: a math quiz with
dynamically increasing difficulty, a memory game in the style of the classic
Simon toy, a pong-style squash game, and a clone of the classic Snake game.