The Linux Development Platform Configuring,Using, and Maintaining a Complete Program Enviroment
Setting up a complete development enviroment using open source tools has always
been a challenging task.

Although all of the development tools are available in the open source,no comprehensive development
enviroment exist as of todayThis book is an effort to enable the reader to set up and use
open source to create such an enviroment.Each chapter of the book is dedicated to a particular
component of the development environment.