Beginning JSON
Programming is not at all a linear path. Often, you find yourself facing a fork in the road. In choosing one
path, you are likely to find, after a period of time, that you go back and travel down the other. While it is
assumed that you are familiar with HTML, CSS, and JavaScript, this book makes no further assumptions
regarding your experience. Therefore, it attempts to provide a thorough explanation for everything you will
read in it.

While JSON is the essence of this book, it is not the sole topic discussed. While that may sound
counterproductive, it is a much-needed requirement. JSON can be devised in isolation, but it would serve
little purpose. What makes JSON so impactful is that it interacts with the many tools of the developer. For
this reason, this book covers a wide range of implementations—from libraries to software.