Learning Linux Shell Scripting
**Who this book is for**
This book is for the readers that are proficient at working with Linux, and who
want to learn about Shell scripting to improve their efficiency and practical skills.
The following are a few examples where we can use skills learned in this book:

•	 Shell scripting is for automating tasks such as taking a periodic backup
•	 Systems administration
•	 Database maintenance and backup
•	 Test processing and report generation
•	 The customization of system initialization

Shell scripts are an essential part of any modern operating system, such as UNIX,
Linux, Windows, and similar. The scripting language or its syntax may vary from
OS to OS; but the fundamental principles remain the same

In this book, you will learn about the very basics of Shell scripting to complex,
customized automation. By the end of this book, you will be able to confidently
use your own Shell scripts for the real-world problems out there. The idea is to be
as practical as possible and give you the look and feel of what real-world scripting
looks like.