GNU/Linux Basics
The objective of this course is to introduce us to the world of GNU/Linux.
It will also provide us with the keys to understanding the philosophy of free
software, teaching us how to use it and to handle it to our taste, giving us
the tools required to move easily in this new world. This document does not
intend to be an essential reference guide for administrators and/or users; there
are already hundreds of manuals, HOWTOS and many other references de-
signed for this purpose, which would occupy millions of pages. Our aim here
is to teach you how to take your first steps in this as yet unexplored field for
many users and administrators and how to approach and solve by yourselves
the problems that you may encounter.