The GNU/Linux Advanced Adminstration By Free Technology Academy (FTA)
The GNU/Linux systems have reached an important level of maturity, allowing to integrate them in almost any kind of work environment,
from a desktop PC to the sever facilities of a big company. In this ebook, the main contents are related with system administration.
You will learn how to install and configure several computer services, and how to optimize and synchronize the resources.

The topics covered in this 500+ page eBook include Linux network, server and data administration, kernel, security, clustering, configuration,
tuning, optimization, migration and coexistence with non-Linux.