Linux: Embedded Development Learning Path
An embedded system is a device with a computer inside that doesn't look like a
computer. Washing machines, televisions, printers, cars, aircraft, and robots are all
controlled by a computer of some sort, and in some cases, more than one. As these
devices become more complex, and as our expectations of the things that we can do
with them expand, the need for a powerful operating system to control them grows.

Linux is only one component of the operating system. Many other components are
needed to create a working system, from basic tools, such as a command shell, to
graphical user interfaces, with web content and communicating with cloud services.
The Linux kernel together with an extensive range of other open source components
allow you to build a system that can function in a wide range of roles.

The Linux kernel is at the heart of a large number of embedded products being
designed today. Over the last 10 years, this operating system has developed from
dominating the server market to being the most used operating system in embedded
systems, even those with real-time requirements.

But at the same time, an embedded Linux product is not only the Linux kernel.
Companies need to build an embedded system over the operating system, and
that's where embedded Linux was finding it difficult to make its place—until
Yocto arrived.

The Yocto Project brings all the benefits of Linux into the development of embedded
systems. It provides a standard build system that allows you to develop embedded
products in a quick, reliable, and controlled way.