ASP.Net Core
ASP.NET Core is the web development framework that comes together with the new .NET Core
and, besides all the new features, also adopts a significantly new approach to web
development.
The first chapter starts by going through the history of Microsoft's web stack to show the
motivations that led to this framework. Later, it moves to more practical matters, like showing
you how to get started with .NET Core and describing the foundations of the framework.