Jump Start MySQL
Preface:
This book is an introduction to the basic concepts of working with a Relational
Database Management System (RDBMS)—specifically, the popular, open source
RDBMS MySQL. Like others Jump series, it aims to give you a head start in your understanding of the chosen technology.
You’ll learn the basics quickly, in a friendly, (hopefully) pain-free way, and have a solid
foundation to continue on in your learning.

Who should read this book:
This book is aimed at those interested in working with data and want to learn how
to use MySQL. To get the most out of some parts of this book, you should have some
previous programming experience, although no specific language is required.