Practical Linux Security Cookbook
When setting up a Linux system, security is supposed to be an important part of all stages.
A good knowledge of the fundamentals of Linux is essential to implementing a good security
policy on the machine.

Practical Linux Security Cookbook will work as a practical guide for administrators and help
them configure a more secure machine.

If you want to learn about Kernel configuration, filesystem security, secure authentication,
network security, and various security tools for Linux, this book is for you.

Linux security is a massive subject and not everything can be covered in just one book. Still,
Practical Linux Security Cookbook will give you a lot of recipes for securing your machine.