Jump Start Responsive Design
This book aims to get you started understanding and using the suite of CSS and
HTML tools for responding to this new world of devices. It will introduce you to
the building blocks that help your pages adapt to different devices, take this a
step further with grids, show you how to make text and images readable on all
devices, and, in the final chapter, cover how to utilize more esoteric device
capabilities.

Who Should Read This Book

This book is for web designers and front-end developers. You’ll need to be
familiar with HTML and CSS, but no previous responsive web design experience
is required. Some JavaScript familiarity is useful for the latter parts of the book.