Mongo DB 3
The purpose of this book is to make you aware of the MongoDB NoSQL database so you can
start working with this database in the fastest way.

This book is intended for software developers who are already working with relational databases
and want to know a bit more about MongoDB. Readers with a good understanding of database
concepts, the Microsoft.NET Framework (most of the examples are written in C#), and JSON
will benefit the most. Some shortcuts have been taken in this book, as the intention is not to go
too deep into the content, but rather to show the various options, possibilities, and concepts.