Angular 2
Angular 2 is a complete rewrite of the Angular library, and is not backwards compatible with
Angular 1 applications. This caused some concern among developers, but the Angular team
wanted to take advantage of many new features that were not around in 2009. Angular 2 is
about making the most of the new browser developments to move forward and create better
applications.