Node.Js
This book is a quick view on Node.js, this new, astonishing, popular, fresh technology
that is spreading through developers all around the world. This is a good introduction to the platform
that inspects various topics from basic to intermediate. It also provides an overview on the
environment around the platform so that after this book, the readers should have a clear idea
about what they can do with Node and how to start using it. To read and fully understand this
book, you need to know JavaScript.