Python Cookbook 3er Edition
Who This Book Is For?

If you need help writing programs in Python 3, or want to update older Python 2 code,
this book is just the ticket. Packed with practical recipes written and tested with Python 3.3,
this unique cookbook is for experienced Python programmers who want to focus on modern tools and idioms.

Inside, you'll find complete recipes for more than a dozen topics, covering the core Python language
as well as tasks common to a wide variety of application domains.
Each recipe contains code samples you can use in your projects right away, along with a discussion
about how and why the solution works.