Oracle PL/SQL
PL/SQL is a combination of SQL along with the procedural features of programing language.
It was developed by Oracle Corporation in the early 90's to enhance the capabilities of SQL.

PL/SQL is one of the three key programming language embedded in the Oracle database
and other advanced RDBMS concepts.