JavaScript Interview Questions
JavaScript is a high level, dynamic, untyped, and interpreted programming language.
It has been standardized in the ECMAScript language specification. Alongside HTML and CSS
it is one of the three essential technologies of the World Wide Web content production; the majority of the
websites employ it and it is supported by all modern web browsers without plug-ins or any kind of other
extension.