The Majesty of Vue.Js 2
Vue is a progressive framework for building user interfaces.Unlike
other monolithic frameworks, Vue is designed from the ground up to be incrementally adoptable.
The core library is focused on the view layer only, and is very easy to pick up and integrate
with other libraries or existing projects.

On the  other hand, Vuew is also perfectly capable of
powering sophisticated Single-Page Application when used in combination with modern tooling
and supporting libraries.