Basic Programmable Logic Controller Principles
Programmable logic controllers (PLC's) permit hardware control devices such
as relays, timers, counters, and drum controllers (sequencers) to be replaced by
programmable solid-state components and programmed instructions. To do so, a
ladders program, consisting of a set of instructions reprensenting the logic to be
followed by the PLC, is developed, entered, and downloaded to the PLC.Once placed in the
Run mode, the PLC follows this logic to interpret the input signals sent to it from input devices
and operate its output devices accordingly.