CSS MASTER Second Edition
This book is for intermediate level CSS developers, as it assumes a fair amount of experience
with html and CSS. Coverage if CSS concepts such as the box model and positioning are included to illuminate
tricky concepts fo the experienced developer, but this coverage is not meant as an introduction for beginners.
Experience with JavaScript is helpful, but not necessary.