Entity Framework Core (EFC)
Entity Framework Core is the brand-new data access technology from Microsoft.
It is a complete rewrite from the "classic" Entity Framework, yet follows the Code First direction.

Entity Framework Core(EFC) builds on Entity Framework Code First, a very popular release that was introduced
with the Code Firts workflow in mind: developers would first create POCO classes to map the concepts
they want to reproduce, and Entity Framework would take care of their persistence in a semi-transparent way.

So, whats is new ? Well, a lot, actually. For one, it builds on the new .NET Core, which is multiplatform.
Yeah, you heard it right:you code will rin on Linux and MAX, in addition to good, old Windows! but that is not all
EFC will feature the capability to connectio to nonrelationals data sources--think NoSQL databases.