Clean Code by Robert C. Martin
Clean Code: A Handbook of Agile Software Craftsmanship is divided into three parts. The first describes the principles, patterns, and practices of writing clean code.
The second part consists of several case studies of increasing complexity. 
Each case study is an exercise in cleaning up code—of transforming a code base that has some problems into one that is sound and efficient. 
The third part is the payoff: a single chapter containing a list of heuristics and “smells” gathered while creating the case studies.
The result is a knowledge base that describes the way we think when we write, read, and clean code.